agent/requirements.txt
bcrypt==4.1.2
blinker==1.4
bullet==2.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
command-not-found==0.3
crit==0.0.1
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
grpcio==1.60.1
grpcio-tools==1.60.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
paramiko==3.4.0
parsys-requests-unixsocket==0.3.1
podman==4.9.0
protobuf==4.25.2
PyByteBuffer==1.0.5
pycairo==1.26.0
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pyxdg==0.28
PyYAML==5.4.1
requests==2.31.0
scp==0.14.5
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tomli==2.0.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
wadllib==1.3.6
zipp==1.0.0
